# client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finance Compliance RAG API",
    "description": "Retrieval-grounded, multi-agent compliance answers over CSSF, DORA, EBA, and NIS2 regulatory text. Advisory only \u2014 not legal advice; every answer stays traceable to a cited source chunk.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Run Query",
        "operationId": "run_query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/stream": {
      "post": {
        "summary": "Run Query Stream",
        "operationId": "run_query_stream_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/{trace_id}": {
      "get": {
        "summary": "Get Query",
        "operationId": "get_query_query__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/{trace_id}/review": {
      "post": {
        "summary": "Submit Review",
        "description": "Human-in-the-loop approve/reject/annotate action (roadmap \u00a73.6) \u2014 can\nonly be filed against a trace_id that actually completed a run, so a\ntypo'd trace_id can't silently create an orphan review record.",
        "operationId": "submit_review_query__trace_id__review_post",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecord"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/{trace_id}/reviews": {
      "get": {
        "summary": "Get Reviews",
        "operationId": "get_reviews_query__trace_id__reviews_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewRecord"
                  },
                  "title": "Response Get Reviews Query  Trace Id  Reviews Get"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentResultResponse": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "agent_name",
          "answer",
          "confidence"
        ],
        "title": "AgentResultResponse"
      },
      "AuditTrail": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "agents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agents"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "token_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "query",
          "model_version",
          "agents",
          "timestamp"
        ],
        "title": "AuditTrail"
      },
      "Citation": {
        "properties": {
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "source_regulation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Regulation"
          },
          "similarity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity Score"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "Citation"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "mlflow_tracking_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Tracking Uri"
          }
        },
        "type": "object",
        "required": [
          "status",
          "model_version",
          "llm_model"
        ],
        "title": "HealthResponse"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Overrides Settings.model_version for this request only."
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AgentResultResponse"
          },
          "summary": {
            "$ref": "#/components/schemas/AgentResultResponse"
          },
          "risk": {
            "$ref": "#/components/schemas/AgentResultResponse"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "audit_trail": {
            "$ref": "#/components/schemas/AuditTrail"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "summary",
          "risk",
          "confidence",
          "audit_trail"
        ],
        "title": "QueryResponse"
      },
      "ReviewRecord": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision"
          },
          "reviewer_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reviewer Name"
          },
          "reviewer_role": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Role"
          },
          "annotation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "reviewer_name",
          "trace_id",
          "timestamp"
        ],
        "title": "ReviewRecord"
      },
      "ReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision"
          },
          "reviewer_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reviewer Name"
          },
          "reviewer_role": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Role"
          },
          "annotation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "reviewer_name"
        ],
        "title": "ReviewRequest",
        "description": "Human-in-the-loop approve/reject/annotate action on a completed run\n(roadmap \u00a73.6) \u2014 doubles as the per-answer compliance sign-off record\ndocs/06 \u00a74.10 calls for. `reviewer_name` is self-reported free text:\nthis app has no auth/user-identity system, a deliberate, documented\nlimitation (see docs/09), not an oversight."
      },
      "TokenUsage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "default": 0
          },
          "embedding_tokens": {
            "type": "integer",
            "title": "Embedding Tokens",
            "default": 0
          }
        },
        "type": "object",
        "title": "TokenUsage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}